+++ crates/holochain/src/conductor/interface/websocket.rs
@@ -556,12 +556,12 @@ pub mod test {
 
         let (dna_file, _, _) =
             SweetDnaFile::unique_from_test_wasms(vec![TestWasm::PostCommitSignal]).await;
-        let app_bundle = app_bundle_from_dnas(&[dna_file.clone()]).await;
+        let app_bundle = app_bundle_from_dnas(&[dna_file.clone()], false).await;
         let request = AdminRequest::InstallApp(Box::new(InstallAppPayload {
             source: AppBundleSource::Bundle(app_bundle),
             agent_key: None,
             installed_app_id: None,
-            membrane_proofs: HashMap::new(),
+            membrane_proofs: HashMap::new().into(),
             network_seed: None,
             #[cfg(feature = "chc")]
             ignore_genesis_failure: false,