+++ crates/holochain/src/core/ribosome/host_fn/accept_countersigning_preflight_request.rs
@@ -158,6 +158,7 @@ pub mod wasm_test {
 
     #[tokio::test(flavor = "multi_thread")]
     #[cfg(feature = "slow_tests")]
+    #[cfg_attr(target_os = "macos", ignore = "flaky on macos")]
     async fn unlock_timeout_session() {
         holochain_trace::test_run();
         let RibosomeTestFixture {
@@ -514,6 +515,7 @@ pub mod wasm_test {
 
     #[tokio::test(flavor = "multi_thread")]
     #[cfg(feature = "slow_tests")]
+    #[cfg_attr(target_os = "macos", ignore = "flaky on macos")]
     async fn lock_chain() {
         use holochain_nonce::fresh_nonce;
         holochain_trace::test_run();
@@ -796,7 +798,7 @@ pub mod wasm_test {
             )
             .await;
 
-        await_consistency(10, [&alice_cell, &bob_cell])
+        await_consistency(60, [&alice_cell, &bob_cell])
             .await
             .unwrap();