+++ crates/holochain/CHANGELOG.md
@@ -10,6 +10,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Adds DPKI support. This is not fully hooked up, so the main implication for this particular implementation is that you must be using the same DPKI implementation as all other nodes on the network that you wish to talk to. If the DPKI version mismatches, you cannot establish connections, and will see so as an error in the logs. This work is in preparation for future work which will make it possible to restore your keys if you lose your device, and to revoke and replace your keys if your device is stolen or compromised.
 - Moved the WASM cache from the data directory to a subdirectory of the data directory named `wasm-cache`. Old content
   won't be removed and WASMs will have to be recompiled into the new cache. #3920
+- App manifest now includes a new `membrane_proofs_deferred: bool` field, which allows the membrane proofs for the app's cells to be provided at a time after installation, allowing the app's UI to guide the process of creating membrane proofs.
+- Adds new `AppStatus::AwaitingMemproofs` to indicate an app which was installed with `MemproofProvisioning::Deferred`
+- Adds new app websocket method `ProvideMemproofs` for use with `MemproofProvisioning::Deferred`
+
+## 0.4.0-dev.6
+
+## 0.4.0-dev.5
+
+- Moved the WASM cache from the data directory to a subdirectory of the data directory named `wasm-cache`. Old content won’t be removed and WASMs will have to be recompiled into the new cache. \#3920
 - Remove deprecated functions `consistency_10s` and `consistency_60s`. Use `await_consistency` instead.
 - Remove deprecated type `SweetEasyInline`. Use `SweetInlineZomes` instead.
 - Remove deprecated methods `SweetInlineZomes::callback` and `SweetInlineZomes::integrity_callback`. Use `SweetInlineZomes::function` and `SweetInlineZomes::integrity_function` instead.