+++ crates/holochain_types/CHANGELOG.md
@@ -8,6 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## \[Unreleased\]
 
 - `AppRoleAssignment::base_cell_id` has become `AppRoleAssignment::base_dna_hash`, because the agent key will always be the same for every cell in the app, so it's redundant to include a CellId which contains the same agent key as the app.
+
+## 0.4.0-dev.6
+
+## 0.4.0-dev.5
+
 - Remove deprecated type `DhtOpLight`. Use `DhtOpLite` instead.
 
 ## 0.4.0-dev.4